@@ -13,52 +13,52 @@
     "@babel/plugin-proposal-function-sent": "^7.2.0",
     "@babel/plugin-proposal-json-strings": "^7.2.0",
     "@babel/plugin-proposal-logical-assignment-operators": "^7.2.0",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.2.0",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.4.3",
     "@babel/plugin-proposal-numeric-separator": "^7.2.0",
     "@babel/plugin-proposal-optional-chaining": "^7.2.0",
     "@babel/plugin-proposal-pipeline-operator": "^7.3.2",
     "@babel/plugin-proposal-throw-expressions": "^7.2.0",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@babel/plugin-syntax-import-meta": "^7.2.0",
-    "@babel/plugin-transform-proto-to-assign": "^7.2.0",
+    "@babel/plugin-transform-proto-to-assign": "^7.4.3",
     "@babel/preset-react": "^7.0.0",
     "@babel/register": "^7.4.0",
-    "@babel/runtime": "^7.4.2",
+    "@babel/runtime": "^7.4.3",
     "@storybook/addon-actions": "^4.1.13",
     "@storybook/react": "^4.1.13",
     "@wdio/browserstack-service": "^5.7.6",
-    "@wdio/cli": "^5.7.6",
+    "@wdio/cli": "^5.7.12",
     "@wdio/dot-reporter": "^5.7.2",
-    "@wdio/local-runner": "^5.7.6",
+    "@wdio/local-runner": "^5.7.9",
     "@wdio/mocha-framework": "^5.7.6",
-    "@wdio/spec-reporter": "^5.7.2",
-    "@wdio/sync": "^5.7.6",
+    "@wdio/spec-reporter": "^5.7.9",
+    "@wdio/sync": "^5.7.9",
     "aigle": "^1.12.0",
     "autobind-decorator": "^2.4.0",
-    "autoprefixer": "^9.5.0",
+    "autoprefixer": "^9.5.1",
     "ava": "^1.4.1",
     "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.5.0",
+    "babel-jest": "^24.7.1",
     "babel-loader": "^8.0.5",
     "chai": "^4.2.0",
     "chalk": "^2.4.2",
     "chromedriver": "^2.46.0",
     "colors": "^1.3.2",
-    "config": "^3.0.1",
+    "config": "^3.1.0",
     "copy-webpack-plugin": "^5.0.2",
     "cross-spawn": "^6.0.5",
     "css-loader": "^2.1.1",
     "dotenv": "^7.0.0",
-    "eslint": "^5.15.3",
+    "eslint": "^5.16.0",
     "eslint-config-react-app": "^3.0.8",
     "eslint-config-standard": "^12.0.0",
     "eslint-config-standard-react": "^7.0.2",
     "eslint-loader": "^2.1.2",
-    "eslint-plugin-flowtype": "^3.4.2",
+    "eslint-plugin-flowtype": "^3.6.0",
     "eslint-plugin-import": "^2.16.0",
     "eslint-plugin-jsx-a11y": "^6.2.1",
     "eslint-plugin-node": "^8.0.1",
-    "eslint-plugin-promise": "^4.0.1",
+    "eslint-plugin-promise": "^4.1.1",
     "eslint-plugin-react": "^7.12.4",
     "eslint-plugin-standard": "^4.0.0",
     "faker": "^4.1.0",
